feat/contextMenu
1. 키보드 이벤트 추가
2. 코드 정리
